Simplified entity to dto mapping throughout and removed obsolete code

+++ piggybank-rest/src/main/java/io/github/joaovicente/piggybank/service/TransactionService.java
@@ -1,14 +1,12 @@
 package io.github.joaovicente.piggybank.service;
 
-import io.github.joaovicente.piggybank.dto.IdResponseDto;
-import io.github.joaovicente.piggybank.dto.TransactionDto;
-import io.github.joaovicente.piggybank.dto.TransactionReadDto;
 import io.github.joaovicente.piggybank.entity.Transaction;
 import io.github.joaovicente.piggybank.repository.TransactionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TransactionService {
@@ -16,30 +14,40 @@ public class TransactionService {
     private KidService kidService;
 
     @Autowired
-    public TransactionService(TransactionRepository transactionRepository, KidService kidService) {
+    public TransactionService(
+            TransactionRepository transactionRepository, KidService kidService) {
         this.transactionRepository = transactionRepository;
         this.kidService = kidService;
     }
 
-    public IdResponseDto createTransaction(TransactionDto requestDto) {
-        Transaction transaction = null;
-        transactionRepository.insert(transaction);
-        IdResponseDto responseDto = IdResponseDto.builder()
-                .id(transaction.getId())
-                .build();
-        return responseDto;
+    public Transaction createTransaction(Transaction transaction) {
+        Transaction insertedTransaction = transactionRepository.insert(transaction);
+        return insertedTransaction;
     }
 
-    public List<TransactionReadDto> getTransactionsByKidId(String kidId)  {
-        List<Transaction> modelTransaction = transactionRepository.findByKidId(kidId);
-        List<TransactionReadDto> dtoTransactions = null;
-        //TODO: Convert transaction model to transaction dto
-        return dtoTransactions;
+    public List<Transaction> getTransactionsByKidId(String kidId)  {
+        List<Transaction> transactionList;
+        if (!kidService.kidExists(kidId))    {
+            throw new KidNotFoundException();
+        }
+        transactionList = transactionRepository.findByKidId(kidId);
+        return transactionList;
     }
 
+
+    public Transaction getTransactionById(String transactionId)  {
+        Optional<Transaction> transaction = transactionRepository.findById(transactionId);
+        if (!transaction.isPresent())    {
+            throw new TransactionNotFoundException();
+        }
+        return transaction.get();
+    }
 
     public void deleteTransaction(String transactionId)   {
-        //TODO: Throw TransactionNotFound if transactionId does not exist
+        Optional<Transaction> transaction = transactionRepository.findById(transactionId);
+        if (!transaction.isPresent())    {
+            throw new TransactionNotFoundException();
+        }
         transactionRepository.deleteById(transactionId);
     }
